refactored, updated, cleaned up
