--- Data/Application.cs
@@ -1,12 +1,0 @@
-﻿using System.Runtime.Serialization;
-
-namespace DataContractDeserialization.Data
-{
-    [DataContract(Namespace = "http://xyz")]
-    public class Application
-    {
-
-        [DataMember]
-        public string Amount;
-    }
-}
