--- Data/ApplicationRequestImpl.cs
@@ -1,14 +1,0 @@
-﻿using System.Runtime.Serialization;
-
-namespace DataContractDeserialization.Data
-{
-    [DataContract(Namespace = "http://xyz")]
-    public class ApplicationRequestImpl
-    {
-        [DataMember]
-        public Application Application;
-
-        [DataMember]
-        public Application_SA[] Application_SAList;
-    }
-}
