--- Data/Application_SA.cs
@@ -1,11 +1,0 @@
-﻿using System.Runtime.Serialization;
-
-namespace DataContractDeserialization.Data
-{
-    [DataContract(Namespace = "http://xyz")]
-    public class Application_SA
-    {
-        [DataMember]
-        public string Security_Key;
-    }
-}
