+++ DataContractDeserialization.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29123.88
+# Visual Studio Version 17
+VisualStudioVersion = 17.13.35825.156 d17.13
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataContractDeserialization", "DataContractDeserialization.csproj", "{EEFF115F-A11F-444C-B675-445BF76EA09F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataContractDeserialization", "src\DataContractDeserialization.csproj", "{EEFF115F-A11F-444C-B675-445BF76EA09F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
