+++ src/DataContractDeserialization.csproj.SdkResolver.1981936763.proj
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <DOTNET_EXPERIMENTAL_HOST_PATH>C:\Program Files\dotnet\dotnet.exe</DOTNET_EXPERIMENTAL_HOST_PATH>
+    <SdkResolverMSBuildTaskHostRuntimeVersion>9.0.2</SdkResolverMSBuildTaskHostRuntimeVersion>
+  </PropertyGroup>
+</Project>
